many fixes

+++ common/src/main/java/net/cristellib/config/Placement.java
@@ -3,6 +3,6 @@ package net.cristellib.config;
 public class Placement {
     public int spacing;
     public int separation;
-    public float frequency;
+    public double frequency;
     public int salt;
 }
